+++ project/src/components/cities-list/cities-list-component.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import { useAppDispatch } from '../../hooks';
-import { setCity } from '../../store/action';
+import { setCity } from '../../store/actions';
 
 type CitiesListComponentProps = {
   citiesList: string[];
