+++ project/src/components/login/login-header.tsx
@@ -1,28 +1,14 @@
 import { Link } from 'react-router-dom';
-import { useAppDispatch } from '../../hooks';
-import { logoutAction } from '../../store/api-actions';
 
 function LoginHeader() {
-  const dispatch = useAppDispatch();
-
-  return (
+  return(
     <nav className="header__nav">
       <ul className="header__nav-list">
         <li className="header__nav-item user">
-          <Link to={'/favorites'} className="header__nav-link header__nav-link--profile">
+          <Link to={'/login'} className="header__nav-link header__nav-link--profile">
             <div className="header__avatar-wrapper user__avatar-wrapper">
             </div>
-            <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-            <span className="header__favorite-count">3</span>
-          </Link>
-        </li>
-        <li className="header__nav-item" onClick={(evt) => {
-          evt.preventDefault();
-          dispatch(logoutAction());
-        }}
-        >
-          <Link to={'/'} className="header__nav-link" >
-            <span className="header__signout">Sign out</span>
+            <span className="header__login">Sign in</span>
           </Link>
         </li>
       </ul>
