+++ project/src/components/login/logout-header.tsx
@@ -1,14 +1,29 @@
 import { Link } from 'react-router-dom';
+import broserHistory from '../../browser-history';
+import { useAppDispatch } from '../../hooks';
+import { logoutAction } from '../../store/api-actions';
 
 function LogoutHeader() {
-  return(
+  const dispatch = useAppDispatch();
+
+  return (
     <nav className="header__nav">
       <ul className="header__nav-list">
         <li className="header__nav-item user">
-          <Link to={'/login'} className="header__nav-link header__nav-link--profile">
+          <Link to={'/favorites'} className="header__nav-link header__nav-link--profile">
             <div className="header__avatar-wrapper user__avatar-wrapper">
             </div>
-            <span className="header__login">Sign in</span>
+            <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
+            <span className="header__favorite-count">3</span>
+          </Link>
+        </li>
+        <li className="header__nav-item" onClick={(evt) => {
+          evt.preventDefault();
+          dispatch(logoutAction());
+        }}
+        >
+          <Link to={broserHistory.location} className="header__nav-link" >
+            <span className="header__signout">Sign out</span>
           </Link>
         </li>
       </ul>
