+++ project/src/pages/login/login-page.tsx
@@ -1,4 +1,5 @@
 import LoginComponent from '../../components/login/login-component';
+import { Link } from 'react-router-dom';
 
 function LoginPage() {
   return(
@@ -7,9 +8,9 @@ function LoginPage() {
         <div className="container">
           <div className="header__wrapper">
             <div className="header__left">
-              <a className="header__logo-link" href="main.html">
+              <Link to={'/'}className="header__logo-link" >
                 <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width={81} height={41} />
-              </a>
+              </Link>
             </div>
           </div>
         </div>
