+++ project/src/pages/room-page/room-page.tsx
@@ -6,7 +6,7 @@ import ReviewsList from '../../components/reviews/reviews-list';
 import MapComponent from '../../components/map-component/map-component';
 import { Comment } from '../../types/comment';
 import { Offer } from '../../types/offer';
-import { getRating } from '../../utils/index';
+import { getRating, setHeader } from '../../utils/index';
 import { useAppSelector } from '../../hooks';
 
 type RoomPageProps = {
@@ -19,6 +19,7 @@ function RoomPage( props: RoomPageProps) {
   const { id } = useParams<{id:string}>();
   const offers = useAppSelector((state) => state.offers);
   const theOffer = offers.find((offer) => offer.id.toString() === id);
+  const isAuthStatus = useAppSelector((state) => state.authorizationStatus);
 
   if (!theOffer) {
     return null;
@@ -34,23 +35,7 @@ function RoomPage( props: RoomPageProps) {
                 <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width={81} height={41} />
               </a>
             </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
+            {setHeader(isAuthStatus)}
           </div>
         </div>
       </header>
